+++ CHANGELOG.md
@@ -24,6 +24,8 @@
 * Bug fix: `distinctOn` and `distinctOnBy` now return a single row if
   zero columns are chosen to be distinct.
 
+* Add `runInsert`/`Update`/`Delete` without underscore
+
 ## 0.7.6.2
 
 Fix ISO 8601 date fomatting.  Thanks to Michal @kozak.